% Auto-generated: do not edit by hand
\name{''DesignableRadioItems}

\alias{''DesignableRadioItems}

\title{DesignableRadioItems component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
''DesignableRadioItems(id=NULL, className=NULL, n_clicks=NULL, name=NULL,
options=NULL, readonly=NULL, value=NULL)
}

\arguments{
\item{id}{Character. }

\item{className}{Character. }

\item{n_clicks}{Numeric. }

\item{name}{Character. }

\item{options}{Unnamed list. }

\item{readonly}{Logical. }

\item{value}{Character | numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

